requirements.txt
pandas>=1.5.0
pyodbc>=4.0.35
python-dotenv>=1.0.0
